OP-342 GCS/ModelView - Update glc_lib to newest 2.1 release (fixes known issues) and update modelview to use this new lib.
git-svn-id: svn://svn.openpilot.org/OpenPilot/trunk@3063 ebee16cc-31ac-478f-84a7-5cbb03baadba

+++ ground/openpilotgcs/src/libs/glc_lib/3rdparty/lib3ds/mesh.c
@@ -570,8 +570,8 @@ lib3ds_mesh_dump(Lib3dsMesh *mesh)
   ASSERT(mesh);
   printf("  %s vertices=%ld faces=%ld\n",
     mesh->name,
-    (long int)mesh->points,
-    (long int)mesh->faces
+    mesh->points,
+    mesh->faces
   );
   printf("  matrix:\n");
   lib3ds_matrix_dump(mesh->matrix);
